Refactor examples and remove dioxus sample
Simplifies the sign-in and sign-up example code, removing interactive flows and environment variable dependencies. Deletes the dioxus-example directory and related assets, configuration, and documentation. Updates Cargo.toml files to reflect these changes and removes unused dependencies and bin targets.

+++ src/models/client_organization_domain_verification.rs
@@ -18,18 +18,13 @@ pub struct ClientOrganizationDomainVerification {
     #[serde(rename = "strategy")]
     pub strategy: String,
     #[serde(rename = "attempts", deserialize_with = "Option::deserialize")]
-    pub attempts: Option<i64>,
+    pub attempts: Option<i32>,
     #[serde(rename = "expire_at", deserialize_with = "Option::deserialize")]
     pub expire_at: Option<i64>,
 }
 
 impl ClientOrganizationDomainVerification {
-    pub fn new(
-        status: String,
-        strategy: String,
-        attempts: Option<i64>,
-        expire_at: Option<i64>,
-    ) -> ClientOrganizationDomainVerification {
+    pub fn new(status: String, strategy: String, attempts: Option<i32>, expire_at: Option<i64>) -> ClientOrganizationDomainVerification {
         ClientOrganizationDomainVerification {
             status,
             strategy,
@@ -38,3 +33,4 @@ impl ClientOrganizationDomainVerification {
         }
     }
 }
+
