+++ src/models/user_settings_username_settings.rs
@@ -14,19 +14,15 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct UserSettingsUsernameSettings {
     #[serde(rename = "min_length", deserialize_with = "Option::deserialize")]
-    pub min_length: Option<i64>,
+    pub min_length: Option<i32>,
     #[serde(rename = "max_length", deserialize_with = "Option::deserialize")]
-    pub max_length: Option<i64>,
+    pub max_length: Option<i32>,
     #[serde(rename = "allow_extended_special_characters")]
     pub allow_extended_special_characters: bool,
 }
 
 impl UserSettingsUsernameSettings {
-    pub fn new(
-        min_length: Option<i64>,
-        max_length: Option<i64>,
-        allow_extended_special_characters: bool,
-    ) -> UserSettingsUsernameSettings {
+    pub fn new(min_length: Option<i32>, max_length: Option<i32>, allow_extended_special_characters: bool) -> UserSettingsUsernameSettings {
         UserSettingsUsernameSettings {
             min_length,
             max_length,
@@ -34,3 +30,4 @@ impl UserSettingsUsernameSettings {
         }
     }
 }
+
